Split github utils in client and auth
